+++ docs/changelog.md
@@ -37,6 +37,7 @@ correctly with default parameters.
  * [bugfix] Ensure motion high score is reset when thumbnails are disabled [#267](https://github.com/Breakthrough/DVR-Scan/issues/267)
  * [bugfix] Fix "stale" frame references being used as thumbnails [#268](https://github.com/Breakthrough/DVR-Scan/issues/268)
  * [bugfix] Ensure comma-separated timecode list persists in logfiles [#265](https://github.com/Breakthrough/DVR-Scan/issues/265)
+ * [bugfix] Allow unknown keys in config file with a warning [#242](https://github.com/Breakthrough/DVR-Scan/issues/242)
 
 ### 1.8.2.1 (2026-05-09)